use one single command with subcommand groups
